V1

+++ src/main/resources/db/migration/V3__create_registration_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE registration (
     id BIGSERIAL PRIMARY KEY,
-    runner_id INTEGER NOT NULL,
-    race_id INTEGER NOT NULL,
+    runner_id BIGSERIAL NOT NULL,
+    race_id BIGSERIAL NOT NULL,
     registration_date DATE NOT NULL,
 
     CONSTRAINT fk_runner
